requirements.txt
scrapy
mysql-connector-python-rf ~= 2.1